Moved trait.
