+++ src/Service/Tontine/ChargeService.php
@@ -7,8 +7,8 @@ use Illuminate\Support\Facades\DB;
 use Siak\Tontine\Model\Charge;
 use Siak\Tontine\Model\Session;
 use Siak\Tontine\Model\Tontine;
-use Siak\Tontine\Service\Events\EventTrait;
 use Siak\Tontine\Service\TenantService;
+use Siak\Tontine\Service\Traits\EventTrait;
 
 class ChargeService
 {
